Frequency

+++ AppProject2.X/Comparator.h
@@ -34,7 +34,7 @@
 #include <xc.h> // include processor files - each processor file is guarded.
 
 void comparatorInit();
-uint64_t compare();
+uint32_t compare();
 void __attribute__((interrupt, no_auto_psv)) _CompInterrupt(void);
 
 // TODO Insert appropriate #include <>
