+++ AppProject2.X/TimeDelay.h
@@ -9,8 +9,7 @@
  * 
  */
 
-
-void Delay_ms(uint16_t time_ms);
+void Delay_ms(uint32_t time_ms);
 void configTimers();
 void configTimerInterrupt();
 void NewClk(unsigned int clkval);
